+++ src/api/Migrations/20230614140202_AddNationalInterest.cs
@@ -9,7 +9,7 @@ public partial class AddNationalInterest : Migration
     protected override void Up(MigrationBuilder migrationBuilder)
     {
         migrationBuilder.AddColumn<bool>(
-            name: "is_national_interest",
+            name: "national_interest",
             schema: "bdms",
             table: "borehole",
             type: "boolean",
@@ -18,9 +18,5 @@ public partial class AddNationalInterest : Migration
 
     protected override void Down(MigrationBuilder migrationBuilder)
     {
-        migrationBuilder.DropColumn(
-            name: "is_national_interest",
-            schema: "bdms",
-            table: "borehole");
     }
 }