+++ src/api/Migrations/BdmsContextModelSnapshot.cs
@@ -97,10 +97,6 @@ namespace BDMS.Migrations
                         .HasColumnType("double precision")
                         .HasColumnName("inclination_direction_bho");
 
-                    b.Property<bool?>("IsNationalInterest")
-                        .HasColumnType("boolean")
-                        .HasColumnName("is_national_interest");
-
                     b.Property<bool?>("IsPublic")
                         .HasColumnType("boolean")
                         .HasColumnName("public_bho");
@@ -144,6 +140,10 @@ namespace BDMS.Migrations
                     b.Property<string>("Municipality")
                         .HasColumnType("text")
                         .HasColumnName("municipality_bho");
+
+                    b.Property<bool?>("NationalInterest")
+                        .HasColumnType("boolean")
+                        .HasColumnName("national_interest");
 
                     b.Property<string>("OriginalName")
                         .HasColumnType("text")