+++ src/api/Models/Borehole.cs
@@ -147,8 +147,8 @@ public class Borehole : IChangeTracking, IIdentifyable
     /// <summary>
     /// Gets or sets the <see cref="Borehole"/>'s national interest.
     /// </summary>
-    [Column("is_national_interest")]
-    public bool? IsNationalInterest { get; set; }
+    [Column("national_interest")]
+    public bool? NationalInterest { get; set; }
 
     /// <summary>
     /// Gets or sets the <see cref="Borehole"/>'s original name.